Move all the blockdevicefactory functions into testtools
